fix(refactor): reorganize temporal into modular packages and update all imports
Move activities into codebase_workflow_run/ and repository_workflow_run/
sub-packages, extract temporal_workflows.py into workflows/ with focused
runner modules, add observability/ for Logfire setup, and introduce utils/
for error enrichment and cancellation helpers. Update all import paths in
worker manager, workflow service, and interceptor.
