+++ unoplat-code-confluence-query-engine/src/unoplat_code_confluence_query_engine/services/observability/__init__.py
@@ -1,0 +1,1 @@
+"""Observability services for telemetry, instrumentation, and tracing wiring."""
