+++ unoplat-code-confluence-query-engine/src/unoplat_code_confluence_query_engine/services/temporal/activities/repository_workflow_run/git_ref_resolution_activity.py
