+++ unoplat-code-confluence-query-engine/src/unoplat_code_confluence_query_engine/services/temporal/workflows/runners/development_workflow_runner.py
@@ -1,0 +1,150 @@
+from __future__ import annotations
+
+from datetime import timedelta
+import traceback
+
+from temporalio import workflow
+
+with workflow.unsafe.imports_passed_through():
+    from loguru import logger
+
+    from unoplat_code_confluence_query_engine.models.output.agents_md_updater_output import (
+        SectionId,
+    )
+    from unoplat_code_confluence_query_engine.models.repository.repository_ruleset_metadata import (
+        CodebaseMetadata,
+    )
+    from unoplat_code_confluence_query_engine.models.runtime.agent_dependencies import (
+        AgentDependencies,
+    )
+    from unoplat_code_confluence_query_engine.models.statistics.agent_usage_statistics import (
+        UsageStatistics,
+    )
+    from unoplat_code_confluence_query_engine.services.repository.engineering_workflow_service import (
+        normalize_engineering_workflow,
+    )
+    from unoplat_code_confluence_query_engine.services.temporal.activities.codebase_workflow_run.engineering_workflow_completion_activity import (
+        EngineeringWorkflowCompletionActivity,
+    )
+    from unoplat_code_confluence_query_engine.services.temporal.agent_assembly.agents.user_prompts.build_user_prompt_development_workflow import (
+        build_development_workflow_prompt,
+    )
+    from unoplat_code_confluence_query_engine.services.temporal.interceptors.agent_workflow_interceptor import (
+        DB_ACTIVITY_RETRY_POLICY,
+    )
+    from unoplat_code_confluence_query_engine.services.temporal.statistics_helpers import (
+        create_zero_usage_statistics,
+        extract_usage_statistics,
+    )
+    from unoplat_code_confluence_query_engine.services.temporal.temporal_agents import (
+        TemporalAgentRegistry,
+        get_cached_usage_limits,
+    )
+    from unoplat_code_confluence_query_engine.services.temporal.utils import (
+        enrich_agent_error_with_model_details,
+        raise_if_temporal_cancellation,
+    )
+    from unoplat_code_confluence_query_engine.services.temporal.workflows.runners.section_updater_runner import (
+        run_section_updater,
+    )
+
+
+async def run_development_workflow_agent(
+    temporal_agents: TemporalAgentRegistry,
+    repository_qualified_name: str,
+    codebase_metadata: CodebaseMetadata,
+    repository_workflow_run_id: str,
+    programming_language_metadata: dict[str, object],
+    results: dict[str, object],
+    agent_stats: list[UsageStatistics],
+    agent_errors: list[dict[str, object]],
+) -> None:
+    """Run the development workflow agent and section updater."""
+    development_workflow_agent = temporal_agents.development_workflow_guide
+    if development_workflow_agent is None:
+        logger.info(
+            "[workflow] development_workflow_guide is disabled, skipping for {}",
+            codebase_metadata.codebase_name,
+        )
+        agent_stats.append(create_zero_usage_statistics())
+        return
+
+    try:
+        logger.info(
+            "[workflow] Running development_workflow_guide for {}",
+            codebase_metadata.codebase_name,
+        )
+        engineering_workflow_deps = AgentDependencies(
+            repository_qualified_name=repository_qualified_name,
+            codebase_metadata=codebase_metadata,
+            repository_workflow_run_id=repository_workflow_run_id,
+            agent_name="development_workflow_guide",
+        )
+        logger.debug("[workflow] Calling development_workflow_guide.run()...")
+        workflow_result = await development_workflow_agent.run(
+            build_development_workflow_prompt(
+                codebase_path=codebase_metadata.codebase_path,
+                programming_language=codebase_metadata.codebase_programming_language,
+                package_manager=codebase_metadata.codebase_package_manager,
+            ),
+            deps=engineering_workflow_deps,
+            usage_limits=get_cached_usage_limits(),
+        )
+        logger.debug("[workflow] development_workflow_guide.run() returned")
+
+        raw_engineering_workflow = workflow_result.output.model_dump(mode="json")
+        normalized_workflow = normalize_engineering_workflow(raw_engineering_workflow)
+        results["engineering_workflow"] = normalized_workflow.model_dump()
+
+        await workflow.execute_activity(
+            EngineeringWorkflowCompletionActivity.emit_engineering_workflow_completion,
+            args=[
+                repository_qualified_name,
+                repository_workflow_run_id,
+                codebase_metadata.codebase_name,
+                codebase_metadata.codebase_programming_language,
+            ],
+            start_to_close_timeout=timedelta(seconds=30),
+            retry_policy=DB_ACTIVITY_RETRY_POLICY,
+        )
+
+        logger.info(
+            "[workflow] development_workflow_guide completed for {}",
+            codebase_metadata.codebase_name,
+        )
+        agent_stats.append(extract_usage_statistics(workflow_result.usage()))
+
+        await run_section_updater(
+            temporal_agents=temporal_agents,
+            section_id=SectionId.ENGINEERING_WORKFLOW,
+            codebase_metadata=codebase_metadata,
+            repository_qualified_name=repository_qualified_name,
+            repository_workflow_run_id=repository_workflow_run_id,
+            programming_language_metadata=programming_language_metadata,
+            section_data=results["engineering_workflow"],
+            agent_stats=agent_stats,
+            agent_errors=agent_errors,
+            updater_runs=results["agents_md_updater_runs"],
+        )
+    except Exception as e:
+        raise_if_temporal_cancellation(e)
+        logger.error(
+            "[workflow] development_workflow_guide failed for {}: {}",
+            codebase_metadata.codebase_name,
+            e,
+        )
+        logger.exception("[workflow] Full traceback:")
+        engineering_error: dict[str, object] = {
+            "agent": "development_workflow_guide",
+            "codebase": codebase_metadata.codebase_name,
+            "error": str(e),
+            "traceback": traceback.format_exc(),
+        }
+        engineering_error = enrich_agent_error_with_model_details(
+            engineering_error,
+            e,
+            "development_workflow_guide",
+            codebase_metadata.codebase_name,
+        )
+        agent_errors.append(engineering_error)
+        agent_stats.append(create_zero_usage_statistics())
